% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_synapse.R
\name{check_public_access}
\alias{check_public_access}
\title{Check public download access}
\usage{
check_public_access(benefactor_id)
}
\arguments{
\item{benefactor_id}{A benefactor id to check; usually a higher-level container, but files can be their own benefactor.}
}
\description{
This returns whether a benefactor id allows public download and view access to the public as
\code{list(READ = bool, DOWNLOAD = bool)}.
}
