% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sankey.R
\name{check_transition}
\alias{check_transition}
\title{Sanity check valid transitions}
\usage{
check_transition(status_data)
}
\arguments{
\item{status_data}{Status data obtained from \code{query_data_status_snapshots}.}
}
\description{
Helper for seeing inconsistent records for status changes...
}
