% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{connect_to_dw}
\alias{connect_to_dw}
\title{Helper for establishing connection}
\usage{
connect_to_dw(config_file)
}
\arguments{
\item{config_file}{YAML config file. See \code{synapseusagereports} package docs for example config file.}
}
\value{
Connection object.
}
\description{
Helper for establishing connection
}
