% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dw_config}
\alias{dw_config}
\title{Write a default config yaml}
\usage{
dw_config(host = "", port = 3306, username = "", password = "")
}
\arguments{
\item{host}{Host URI.}

\item{port}{Numeric port; defaults to 3306.}

\item{username}{Username.}

\item{password}{Password.}
}
\description{
Create a \code{config.yml} to store datawarehouse connection parameters if you don't already one.
}
