% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{encrypt_code_file}
\alias{encrypt_code_file}
\title{Encrypt and save code file}
\usage{
encrypt_code_file(
  code_df,
  code_file = "codes.csv",
  key_rds = "usage_report_key.rds"
)
}
\arguments{
\item{code_df}{A \code{data.frame} with cols \code{userId} and \code{code} to encrypt.}

\item{code_file}{Name of file to encrypt to, should have \code{csv} extension.}

\item{key_rds}{Encryption key.}
}
\description{
See also inverse helper \code{read_code_file}.
}
