% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_circular.R
\name{fig_class_count}
\alias{fig_class_count}
\title{Hierarchical circle packing}
\usage{
fig_class_count(nodes, edges, seed = 42, background = "black")
}
\arguments{
\item{nodes}{Nodes data. See \code{simd_class_count} to get example.}

\item{edges}{Edges data. See \code{simd_class_count} to get example.}

\item{seed}{Optional, seed for layout.}

\item{background}{Optional, background color.}
}
\description{
This is similar to a treemap and is useful for representing proportions of hierarchical entities (ontology classes and subclasses).
}
