% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{flowchart_component_template}
\alias{flowchart_component_template}
\title{Add additional component to flowchart}
\usage{
flowchart_component_template(
  qualify,
  accepted,
  in_preparation,
  component = "cBP",
  connector = "Reprocessed->cBP_qualify;"
)
}
\arguments{
\item{qualify}{Number for qualification of something.}

\item{accepted}{Number for accepted.}

\item{in_preparation}{Summary number for in preparation.}

\item{component}{Prefix for component id, defaults to "cBP" for cBioPortal}

\item{connector}{Connecting edge if adding this to an existing plot; use NULL for stand-alone plot.}
}
\description{
Add additional component to flowchart
}
