% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{generate_code}
\alias{generate_code}
\title{Internal helper to return a vector of random codes of size needed}
\usage{
generate_code(x, exclude)
}
\arguments{
\item{x}{Vector of ids that need coding.}

\item{exclude}{Vector of codes already in use.}
}
\description{
Internal helper to return a vector of random codes of size needed
}
