% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sankey.R
\name{growth_projects_data_available}
\alias{growth_projects_data_available}
\title{Get change in projects with data available}
\usage{
growth_projects_data_available(
  status_data,
  qualifying = c("Partially Available", "Available")
)
}
\arguments{
\item{status_data}{Status data obtained from \code{query_data_status_snapshots}.}

\item{qualifying}{Qualifying statuses.}
}
\description{
Get change in projects with data available
}
