% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{increasing_stat}
\alias{increasing_stat}
\title{Stat box element}
\usage{
increasing_stat(value, header = NULL, description = NULL)
}
\arguments{
\item{value}{Stat value, e.g. "+1,243" or "+30\%".}

\item{header}{Title for stat value.}

\item{description}{Description or stat value (1-2 sentences).}
}
\description{
Similar element to \code{summary_stat}, but with icon instead of description.
}
