% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{info_format}
\alias{info_format}
\title{Apply all formatting for infographic values}
\usage{
info_format(value, percent = FALSE, delta = TRUE)
}
\arguments{
\item{value}{Numeric value.}

\item{percent}{Add "\%" suffix.}

\item{delta}{Whether to prepend "+" or "-" if this is a delta value that needs the emphasis.}
}
\description{
Convenience function to apply a number of formatting options at once.
}
