% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{notes}
\alias{notes}
\title{Notes block element}
\usage{
notes(header, description)
}
\arguments{
\item{header}{Title for stat value.}

\item{description}{Description or stat value (1-2 sentences).}
}
\description{
Notes block element
}
