% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_bar_data_segment}
\alias{plot_bar_data_segment}
\title{Bar plot of downloads by data type}
\usage{
plot_bar_data_segment(data, title)
}
\arguments{
\item{data}{The data.}

\item{title}{Title of plot.}
}
\description{
Bar plot of downloads by data type
}
