% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_bar_visitors}
\alias{plot_bar_visitors}
\title{Plot visitors}
\usage{
plot_bar_visitors(data, pictogram = FALSE)
}
\arguments{
\item{data}{Data from Google Analytics, something like \code{ga_data}.}

\item{pictogram}{Does additional fun stuff with user pictograms, defaults to \code{FALSE}.}
}
\description{
Plot unique visitors by project, meant to complement \code{plot_dot_pageviews}.
}
