% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_bipartite}
\alias{plot_bipartite}
\title{Bipartite network connecting users and projects}
\usage{
plot_bipartite(data)
}
\arguments{
\item{data}{The data.}
}
\description{
Bipartite network connecting users and projects
}
