% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_bar.R
\name{plot_col_files_available}
\alias{plot_col_files_available}
\title{Bar plot comparing files available}
\usage{
plot_col_files_available(data, palette = c("#af316c", "#125E81"))
}
\arguments{
\item{data}{The data.}

\item{palette}{Color palette.}
}
\description{
This expects a summary of files available by month.
}
