% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_col_pageview_}
\alias{plot_col_pageview_}
\title{Column plot comparing pageviews for projects with released vs unreleased data}
\usage{
plot_col_pageview_(data)
}
\arguments{
\item{data}{The data.}
}
\description{
Column plot comparing pageviews for projects with released vs unreleased data
}
