% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_bar.R
\name{plot_data_segment}
\alias{plot_data_segment}
\title{Bar plot of downloads by data attribute}
\usage{
plot_data_segment(data, palette = NULL)
}
\arguments{
\item{data}{Data in long format for plot. Expects cols \code{attribute} with name of attribute,
\code{value} for the attribute value, and \code{proportion} for the share of that value in the overall data. See details.}

\item{palette}{Optional, manual color palette for customized or consistent fill of the attribute values.
The default palette handles up to}
}
\description{
Convenience proportion plot for annotation attribute.
See either \code{simd_resourcetype_breakdown} or \code{simd_assay_breakdown} for example data.
See also \code{plot_status} for very nearly the same plot method with different data.
}
