% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{plot_donut_file_tiers}
\alias{plot_donut_file_tiers}
\title{Donut of file tiers}
\usage{
plot_donut_file_tiers(data, palette)
}
\arguments{
\item{data}{Table with columns \code{Tier} and \code{Count}.}

\item{palette}{Needs two colors for tiers shown by default. See details.}
}
\description{
Plot the restricted files as a proportion of total public files.
Original data can come from \code{query_file_snapshot}.
The tiers are \code{Public - Unrestricted} and \code{Public - Restricted}.
}
