% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_ga.R
\name{plot_dot_pageviews}
\alias{plot_dot_pageviews}
\title{Dot plot comparing pageviews for projects with released vs unreleased data}
\usage{
plot_dot_pageviews(data)
}
\arguments{
\item{data}{The data.}
}
\description{
Dot plot comparing pageviews for projects with released vs unreleased data
}
