% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_downloads_datetime}
\alias{plot_downloads_datetime}
\title{Column plot of downloads over time}
\usage{
plot_downloads_datetime(data, fill = "project", palette)
}
\arguments{
\item{data}{The data.}

\item{fill}{Variable to fill by, defaults to "project".}

\item{palette}{Color palette.}
}
\description{
Bar of downloads over time (day), grouped by project or Sage (NF-OSI) vs. regular users
Grouping by project becomes problematic when in the future there are 8+ projects, making this hard to read
}
