% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_lollipop_download_by_project}
\alias{plot_lollipop_download_by_project}
\title{Lollipop plot of downloads by project.
Expects a \code{data.frame} with \code{project} and \code{downloads}.}
\usage{
plot_lollipop_download_by_project(data, palette)
}
\arguments{
\item{data}{The data.}

\item{palette}{Color palette.}
}
\description{
Lollipop plot of downloads by project.
Expects a \code{data.frame} with \code{project} and \code{downloads}.
}
