% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_circular.R
\name{plot_pie}
\alias{plot_pie}
\title{Pie chart}
\usage{
plot_pie(data)
}
\arguments{
\item{data}{Data, expects a column 'status'.}
}
\description{
A very generic pie chart for 'status'. See also \code{plot_donut_file_tiers}.
}
