% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{plot_project_flowchart_template}
\alias{plot_project_flowchart_template}
\title{Project breakdown flowchart template}
\usage{
plot_project_flowchart_template(
  total,
  available,
  reprocessable,
  unreleased,
  under_embargo,
  pending,
  not_expected
)
}
\arguments{
\item{total}{Total number.}

\item{available}{Number available from total.}

\item{reprocessable}{Number considered reprocessable, portion of available.}

\item{unreleased}{Number of unreleased from total.}

\item{under_embargo}{Number under embargo.}

\item{pending}{Number where data is still being generated.}

\item{not_expected}{Number where data not expected.}
}
\description{
Generate a mermaid.js flowchart breaking down projects into groupings based on data status.
Can use as a starting template and substitute status labels depending on the DCC.
}
