% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{plot_sankey_status}
\alias{plot_sankey_status}
\title{Sankey status transitions}
\usage{
plot_sankey_status(data, palette = data_status_palette)
}
\arguments{
\item{data}{The data.}

\item{palette}{Color palette.}
}
\description{
Sankey status transitions
}
