% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_ga.R
\name{plot_scatter_pageviews_visitors}
\alias{plot_scatter_pageviews_visitors}
\title{Plot scatter of pageviews x visitors}
\usage{
plot_scatter_pageviews_visitors(data, cutoff = 10)
}
\arguments{
\item{data}{Data from Google Analytics, something like \code{ga_data}.}

\item{cutoff}{Cutoff based on unique visitors; projects above this have labels that help highlight them as more popular.}
}
\description{
This is a scatter plot that presents pageviews and visitors together instead of
separately in \code{plot_dot_pageviews} and \code{plot_bar_visitors}. Pageviews (\code{sum_pageviews})
are represented on Y-axis and as point size; visitors (\code{max_users}) are represented on X-axis.
Points are colored by their release group, which should be in column \code{data_release_group}.
A threshold value selectively adds \code{projectId} labels depending on \code{cutoff}.
}
