% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_bar.R
\name{plot_status}
\alias{plot_status}
\title{Project and data status grouped bar plot}
\usage{
plot_status(data, palette = data_status_palette())
}
\arguments{
\item{data}{Data table, expects \code{studyStatus} and \code{dataStatus} columns.}

\item{palette}{Palette for fill colors by \code{dataStatus}. Defaults are provided.}
}
\description{
Project and data status grouped bar plot
}
