% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{project_flowchart_template}
\alias{project_flowchart_template}
\title{Project breakdown flowchart template}
\usage{
project_flowchart_template(
  total,
  available,
  reprocessed,
  unreleased,
  under_embargo,
  pending,
  not_expected,
  orientation = c("TD", "LR")
)
}
\arguments{
\item{total}{Total number.}

\item{available}{Number available from total.}

\item{reprocessed}{Number reprocessed, which can come from available or under embargo.}

\item{unreleased}{Number of unreleased from total.}

\item{under_embargo}{Number under embargo.}

\item{pending}{Number where data is still being generated.}

\item{not_expected}{Number where data not expected.}

\item{orientation}{Chart orientation (defaults to "TD", top-down).}
}
\description{
Generate a mermaid.js flowchart breaking down projects into groupings based on data status.
Can use as a starting point depending on the DCC.
}
