% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{query_data_by_funding_agency}
\alias{query_data_by_funding_agency}
\title{Query wrapper to get data warehouse data}
\usage{
query_data_by_funding_agency(
  con = NULL,
  config_file = NULL,
  start_date,
  end_date,
  fundingAgency = "NTAP",
  data_status = c("Available", "Partially Available", "Under Embargo", "None"),
  table = "syn16787123",
  save = TRUE,
  disconnect = TRUE
)
}
\arguments{
\item{con}{Connection object. If not given, \code{config_file} should be given.}

\item{config_file}{YAML config file. See \code{synapseusagereports} package docs for example config file.}

\item{start_date}{The start data of the report period -- should be first day of some month.}

\item{end_date}{End date of report period, should be last day of the month six months from \code{start_date}.}

\item{fundingAgency}{An NF funding agency to query for.}

\item{data_status}{Character vector values for data status, defaults to all NF status values.}

\item{table}{Reference studies table or whichever table that has information on studies and funding agency.}

\item{save}{Whether to save a copy of study records to working directory.}

\item{disconnect}{Whether to disconnect after query. Default \code{TRUE}}
}
\description{
Requires devtools::install_github("Sage-Bionetworks/synapseusagereports")
This is a wrapper building on synapseusagereports to download NF-relevant data,
which saves data for each project as a separate .csv within folders organized by report type.
}
