% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_synapse.R
\name{query_data_status_snapshots}
\alias{query_data_status_snapshots}
\title{Query data status snapshots}
\usage{
query_data_status_snapshots(
  vRange,
  fundingAgency = "NTAP",
  ref = "syn16787123"
)
}
\arguments{
\item{vRange}{Start and end of versions to look at, e.g. version #1 to version #10.}

\item{fundingAgency}{An NF funding agency to query for.}

\item{ref}{Synapse table to use for query.}
}
\description{
There are two ways to get the history of data status changes for studies:
\enumerate{
\item Download the weekly snapshots of the portal study table.
\item Look up the permissions history of the \verb{Raw Data} folder (this is not possible).
Both are overly complicated workarounds because data status changes are not recorded
with dates in an official manner. Because the second is not possible anyway,
this implements the first to create a data object representing transition states
for the specific funder (default: NTAP) projects
Once there are better data/methods for data status changes, this can be deprecated.
}
}
