% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_synapse.R
\name{query_file_snapshot}
\alias{query_file_snapshot}
\title{Query available files within the first and last month of report period}
\usage{
query_file_snapshot(
  con,
  config_file = NULL,
  start_date,
  end_date,
  start_id_file,
  end_id_file
)
}
\arguments{
\item{con}{Connection object. If not given, \code{config_file} should be given.}

\item{config_file}{YAML config file. See \code{synapseusagereports} package docs for example config file.}

\item{start_date}{The start data of the report period -- should be first day of some month.}

\item{end_date}{End date of report period, should be last day of the month six months from \code{start_date}.}

\item{start_id_file}{Text files of all study ids with released data at start of report period.}

\item{end_id_file}{Text files of all study ids with released data at end of report period.}
}
\description{
The first file snapshot scope (first month of report period) includes unique files for the projects listed in \code{start_id_file}.
The second file snapshot scope (last month of report period) includes unique files for projects listed in \code{end_id_file}.
Usually, there are more projects in the \code{end_id_file} because of new project releases during the report period;
and unless something is wrong, all projects in \code{start} should be in \code{end} (there should be no "un-release" phenomenon).
For each snapshot, we filter for "available" files, which are either public or controlled access, and calculate a simple delta.
}
