% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ga.R
\name{query_ga}
\alias{query_ga}
\title{Convenience wrapper for Google Analytics}
\usage{
query_ga(projects, date_range)
}
\arguments{
\item{projects}{List of project syn ids, used for pagepath filters.}

\item{date_range}{Start and end date.}
}
\description{
Wrapper to query for pageviews and users.
Make sure already authenticated with \code{googleAnalyticsR::ga_auth()}.
}
