% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ga.R
\name{query_ga_project_stats}
\alias{query_ga_project_stats}
\title{Query project stats with GA4}
\usage{
query_ga_project_stats(projects, date_range, ga_property_id = 311611973)
}
\arguments{
\item{projects}{Character vector of one or more project syn ids.}

\item{date_range}{Start and end date.}

\item{ga_property_id}{Google Analytics property id.}
}
\description{
Convenience wrapper to query for selected project stats: pageviews and users.
Result is a list of tables matching number of projects; each table breaks down the stats for the pages (wiki, files, etc.) of that project.
}
\details{
Make sure already authenticated with \code{googleAnalyticsR::ga_auth()}.
}
