% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{query_study_ids}
\alias{query_study_ids}
\title{Helper for looking up studies}
\usage{
query_study_ids(fundingAgency, dataStatus, table = "syn52694652", save = FALSE)
}
\arguments{
\item{fundingAgency}{A single funder name, e.g. "NTAP".}

\item{dataStatus}{Data status(es), e.g. c("Available", "Partially Available")}

\item{table}{Synapse id of table to use.}

\item{save}{Whether to write data to csv. Default \code{FALSE}.}
}
\description{
Studies can be looked up in a Synapse table \emph{or} Snowflake.
This queries studies by funding agency and data status using Portal - Studies table.
Note: If using a portal table, do not need to log in with \code{synapser} first as data is anonymous access.
}
