% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{read_code_file}
\alias{read_code_file}
\title{Decrypt and read code file}
\usage{
read_code_file(code_file = "codes.csv", key_rds = "usage_report_key.rds")
}
\arguments{
\item{code_file}{Name of file to read from, should have \code{csv} extension.}

\item{key_rds}{Encryption key.}
}
\description{
See also inverse helper \code{encrypt_code_file}.
}
