% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{store_snapshot_data}
\alias{store_snapshot_data}
\title{Store deidentified snapshot data in Synapse}
\usage{
store_snapshot_data(name, data, project = "syn39770191")
}
\arguments{
\item{name}{Name of the table.}

\item{data}{Table data.}

\item{project}{Synapse project in which to store; a default location is used.}
}
\description{
Convenience function to store the compiled snapshot after \code{to_identified_export}.
At some point this may not be needed.
}
