% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{summary_stat}
\alias{summary_stat}
\title{Summary stat block element}
\usage{
summary_stat(
  value,
  header = NULL,
  description = NULL,
  format = TRUE,
  delta = FALSE,
  percent = FALSE
)
}
\arguments{
\item{value}{Stat value.}

\item{header}{Title for stat value.}

\item{description}{Description or stat value (1-2 sentences).}

\item{format}{Whether to format numeric values for commas for display. Default TRUE.}

\item{delta}{Whether to prefix positive numbers with "+" for delta values.}

\item{percent}{Whether this value is a percent.}
}
\description{
Summary stat block element
}
