% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{to_deidentified_export}
\alias{to_deidentified_export}
\title{Transform to deidentified data}
\usage{
to_deidentified_export(
  data_dir,
  code_file = "codes.csv",
  key_rds = "usage_report_key.rds"
)
}
\arguments{
\item{data_dir}{Output files from data warehouse query that contains real synapse user ids.}

\item{code_file}{A csv mapping the random ids (\code{code} column) to real synapse user ids (\code{userId} column).}

\item{key}{Key for decrypting \code{code_file}, defaults to using value of env var \code{USAGE_REPORT_KEY}.
The key is in LastPass.}

\item{synproject}{Id of parent project.}
}
\description{
This processes the data warehouse outputs so that data is deidentified and can be stored
in a private Synapse project. All downstream wranlign rely on this intermediate.
Because of the limits on query intervals, it may be hard to get this data for re-analysis otherwise.
Note that any Sage users use negative integers and have to be hard-coded before using this function.
}
