% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sankey.R
\name{to_sankey_data}
\alias{to_sankey_data}
\title{Massage to Sankey plot data}
\usage{
to_sankey_data(status_data, selected)
}
\arguments{
\item{status_data}{Status data obtained from \code{query_data_status_snapshots}.}

\item{selected}{Timepoints \emph{other} than first and last, which is already included by default.}
}
\description{
Massage to Sankey plot data
}
