% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{connect_to_dw}
\alias{connect_to_dw}
\title{Helper for establishing connection}
\usage{
connect_to_dw(config_file)
}
\arguments{
\item{config_file}{YAML config storing db connection creds,
see \href{https://github.com/Sage-Bionetworks/synapseusagereports/blob/master/example-db-config.yml}{example}.}
}
\value{
Connection object.
}
\description{
Helper for establishing connection
}
