% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{plot_bipartite}
\alias{plot_bipartite}
\title{Bipartite network connecting users and projects}
\usage{
plot_bipartite(data, hide_user_id = TRUE)
}
\arguments{
\item{data}{The data.}

\item{hide_user_id}{Don't label user nodes with their ids. See details.}
}
\description{
User ids are supposed to be random numbers (e.g. 6, 19, 40) and not actual user ids,
but to keep the display cleaner when there is potentially \emph{a lot} of users, the ids are hidden by default.
Also, having numbers can be confusing (apparent by questions like: "Does '19' mean the number of users?" -- no, that's just the id).
}
