% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_data_segment}
\alias{plot_data_segment}
\title{Bar plot of downloads by data type}
\usage{
plot_data_segment(
  rdata,
  adata,
  assay_palette,
  resource_palette = resource_type_palette(),
  donut = FALSE
)
}
\arguments{
\item{rdata}{Resource type annotation data.}

\item{adata}{Assay type annotation data.}

\item{donut}{Display as donut/sunburst-like chart.}
}
\description{
Bar plot of downloads by data type
}
