% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_data_segment}
\alias{plot_data_segment}
\title{Bar plot of downloads by data attributes}
\usage{
plot_data_segment(
  rdata,
  adata,
  assay_palette,
  resource_palette = resource_type_palette(),
  donut = FALSE
)
}
\arguments{
\item{rdata}{Resource type annotation data.}

\item{adata}{Assay type annotation data.}

\item{assay_palette}{Color palette for assays,
which might need more customization since assay annotations can be somewhat project-specific.}

\item{resource_palette}{Color palette for resources -- default uses \code{resource_type_palette()},
but provide a custom palette if annotations are project-specific.}

\item{donut}{Display as donut/sunburst-like chart.}
}
\description{
This expects summary breakdown for core annotation attributes, e.g. a table with cols
\code{resourceType}, \code{proportion} and a table with cols \code{assay}, \code{proportion}.
}
