% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs_main.R
\name{plot_downloads_datetime}
\alias{plot_downloads_datetime}
\title{Bar plot of downloads over time}
\usage{
plot_downloads_datetime(data, fill = "project", palette)
}
\arguments{
\item{data}{The data.}

\item{fill}{Variable to fill by, defaults to "project".}

\item{palette}{Color palette.}
}
\description{
Bar of downloads over time, grouped by project or Sage (NF-OSI) vs. regular users.
}
\details{
Some tips:
\enumerate{
\item Scaling: \code{date} data is binned into "2 weeks" during plotting by default,
but the plot x-axis can easily be (re)scaled if needed after examining the result figure, e.g.:
\code{p <- plot_downloads_datetime(data)}
\code{p + ggplot2::scale_x_date(date_breaks = "2 months")}
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{Depending on the download magnitude, y may preferably use log scaling:
`p + ggplot2::scale_y_continuous(trans='log2')`

2. Grouping by project can be problematic when there are 8+ projects, making this hard to read.
}\if{html}{\out{</div>}}
}
