% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{plot_scatter_pageviews_visitors}
\alias{plot_scatter_pageviews_visitors}
\title{Plot scatter of pageviews x visitors}
\usage{
plot_scatter_pageviews_visitors(data, cutoff = 10)
}
\arguments{
\item{data}{Data from Google Analytics, something like \code{ga_data}.}

\item{cutoff}{Cutoff based on visitor number; projects above this have labels that help highlight them as more popular.}
}
\description{
This is an alternative that presents pagesviews and visitors together instead of
separately in \link{`plot_dot_pageviews`} and \link{`plot_bar_visitors`}.
}
