% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{project_flowchart_template}
\alias{project_flowchart_template}
\title{Project breakdown flowchart template}
\usage{
project_flowchart_template(
  total,
  available,
  reprocessed,
  unreleased,
  under_embargo,
  pending,
  not_expected,
  direction = c("TD", "LR")
)
}
\arguments{
\item{total}{Total number.}

\item{available}{Number available from total.}

\item{reprocessed}{Number reprocessed, which can come from available or under embargo.}

\item{unreleased}{Number of unreleased from total.}

\item{under_embargo}{Number under embargo.}

\item{pending}{Number where data is still being generated.}

\item{not_expected}{Number where data not expected.}

\item{direction}{Chart orientation (defaults to "TD" top-down, alternative is "LR" left-right).}
}
\description{
Generate a GraphViz flowchart breaking down projects into groupings based on data status.
Intended for use as a starting point.
}
