% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_data_by_funding_agency}
\alias{query_data_by_funding_agency}
\title{Query wrapper to get data warehouse data}
\usage{
query_data_by_funding_agency(
  con = NULL,
  config_file = NULL,
  start_date,
  end_date,
  fundingAgency = "NTAP",
  all = TRUE
)
}
\arguments{
\item{con}{Connection object. If not given, \code{config_file} should be given.}

\item{config_file}{YAML config file. See \code{synapseusagereports} package docs for example config file.}

\item{fundingAgency}{An NF funding agency to query for.}

\item{all}{Download for all projects associated with \code{fundingAgency}.}
}
\description{
Requires devtools::install_github("Sage-Bionetworks/synapseusagereports")
This is a wrapper building on synapseusagereports to download NF-relevant data,
which saves data for each project as a separate .csv within folders organized by report type.
}
