% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{query_data_by_funding_agency}
\alias{query_data_by_funding_agency}
\title{Query wrapper to get data warehouse data}
\usage{
query_data_by_funding_agency(
  con,
  start_date,
  end_date = NULL,
  fundingAgency = "NTAP",
  data_status = c("Available", "Partially Available", "Under Embargo", "None"),
  table = "syn16787123",
  save = TRUE,
  disconnect = TRUE
)
}
\arguments{
\item{con}{Connection object; use \code{connect_to_dw} to connect.}

\item{start_date}{Start date of the report range in YYYY-MM-DD format -- should be first day of some month, e.g. "2022-03-01".}

\item{end_date}{End date of report range, defaults to six months from \code{start_date}, e.g., "2022-09-01",
but pass in manually if using a different report interval (though can't exceed six months)}

\item{fundingAgency}{An NF funding agency to query for.}

\item{data_status}{Character vector values for data status, defaults to all NF status values.}

\item{table}{Reference studies table or whichever table that has information on studies and funding agency.}

\item{save}{Whether to save a copy of study records to working directory.}

\item{disconnect}{Whether to disconnect after query. Default \code{TRUE}.}
}
\description{
Requires \code{devtools::install_github("Sage-Bionetworks/synapseusagereports")}.
This is a wrapper building on queries refined in synapseusagereports to download NF-relevant data;
for details, see \href{https://github.com/Sage-Bionetworks/synapseusagereports/blob/master/R/lib.R#L13}{code}.
Data is saved for each project as a separate \code{.csv} within folders organized by report type.
}
