% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{query_file_snapshot}
\alias{query_file_snapshot}
\title{Query file snapshot for settings}
\usage{
query_file_snapshot(con, end_date, project_ids)
}
\arguments{
\item{con}{Connection object; use \code{connect_to_dw} to connect.}

\item{end_date}{Latest date to consider for file snapshots, corresponding to the report end date.}

\item{project_ids}{Vector of project ids without the "syn" prefix.}
}
\description{
The intent is to give a general sense of what files are actually visible and
how easy they are to access (based on restrictions in place, which could of course slow down downloads/usage).
Implementation uses \code{NODE_SNAPSHOT}, which has snapshots files when they are created/modified.
This table is used to check the public/controlled/restricted flags:
\itemize{
\item \code{IS_PUBLIC} : Can public see entity?
\item \code{IS_CONTROLLED} : Is entity under Tier 3 access?
\item \code{IS_RESTRICTED} : Is entity under Tier 2 access?
}
}
\details{
More definitions...
\itemize{
\item \verb{Tier 1} : User agrees to a second EULA specific to certain data layers.
\item \verb{Tier 2} : (Tier 1) + User agrees to a second EULA specific to certain data layers.
\item \verb{Tier 3} : (Tier 1) + (Tier 2) + User access must be requested/approved through ACT.
}
}
