% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dw.R
\name{query_study_ids}
\alias{query_study_ids}
\title{Helper for looking up studies}
\usage{
query_study_ids(fundingAgency, data_status, table, save)
}
\arguments{
\item{fundingAgency}{An NF funding agency to query for.}

\item{data_status}{Character vector values for data status, defaults to all NF status values.}

\item{table}{Reference studies table or whichever table that has information on studies and funding agency.}

\item{save}{Whether to save a copy of study records to working directory.}
}
\description{
Query studies by funding agency and data status.
}
