% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sankey.R
\name{sum_data_status_changes}
\alias{sum_data_status_changes}
\title{Auto-select most interesting timepoints and make data for sankey
Sum data status changes}
\usage{
sum_data_status_changes(status_data)
}
\arguments{
\item{status_data}{Status data obtained from \code{query_data_status_snapshots}.}
}
\description{
Helper to construct a table of most prominent transition points to help select which ones to use.
}
