% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_figs.R
\name{summary_stat}
\alias{summary_stat}
\title{Summary stat block element}
\usage{
summary_stat(value, header = NULL, description = NULL)
}
\arguments{
\item{value}{Stat value.}

\item{header}{Title for stat value.}

\item{description}{Description or stat value (1-2 sentences).}
}
\description{
Summary stat block element
}
