% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{to_deidentified_export}
\alias{to_deidentified_export}
\title{Transform to deidentified data}
\usage{
to_deidentified_export(
  data_dir,
  code_file = "codes.csv",
  key_rds = "usage_report_key.rds"
)
}
\arguments{
\item{data_dir}{Output files from data warehouse query that contains real synapse user ids.}

\item{code_file}{A csv mapping the random ids (\code{code} column) to real synapse user ids (\code{userId} column).
This is expected to be encrypted.}

\item{key_rds}{Key file for decrypting \code{code_file}.}
}
\description{
This processes the data warehouse outputs so that data is deidentified and can be stored
in a private Synapse project. All downstream wrangling/plotting rely on this intermediate.
Storing this data is important because of the limits on availability (i.e. data only available for six months prior),
so make sure to store data if there's any chance of later additional analysis or comparison later.
}
\details{
For the recoding itself, DCC Sage users use negative integers and have to be hard-coded before using this function;
this allows easy filtering of DCC users later on with simple \code{userId > 0} clause.
}
