% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sankey.R
\name{to_sankey_data}
\alias{to_sankey_data}
\title{Selected should be timepoints \emph{other} than first and last}
\usage{
to_sankey_data(status_data, selected)
}
\description{
Selected should be timepoints \emph{other} than first and last
}
